backend-new/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
django-environ==0.11.2
psycopg2-binary==2.9.7
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
stripe==7.8.0
twilio==8.10.0
sendgrid==6.10.0
django-debug-toolbar==4.2.0
django-silk==5.0.4
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
django-ratelimit==4.1.0
django-cleanup==8.0.0
django-imagekit==5.0.0
django-taggit==4.0.0
django-mptt==0.15.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==7.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0